Add test for the fillFUV function #42
The fillFUV tests whether the FUV is filled according to the PUV and the PUM.
The current tests pass only if we commit the current changes of the pull request i.e.
setting FUV[i] to true as a default value and letting j go all the way to 15.
Furthermore I changed the attributes of the Main class and its methods to public so
that we have access to them from the Tests class.